ai-features/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq>=0.13.0
pydantic==2.9.2
python-dotenv==1.0.1
firecrawl-py>=1.5.0
psycopg2-binary>=2.9.9
requests>=2.32.0
